fix(cpp): avoid signed/unsigned comparison in geohash precision loop
geohash.length() is size_t; comparing it directly against the int
precision parameter tripped a sign-compare warning under stricter
build flags.

+++ cpp/HybridNitroLocationComplexLogicsCalculation.cpp
@@ -170,7 +170,7 @@ namespace margelo::nitro::nitrolocationtracking {
       int ch = 0;
       std::string geohash = "";
 
-      while (geohash.length() < precision) {
+      while (static_cast<int>(geohash.length()) < precision) {
           if (is_even) {
               double mid = (lon[0] + lon[1]) / 2;
               if (longitude > mid) {
